2.12 Making Camera follow the player
There are several ways to to this. The easiest method is to make the camera a child of the player, therefore wherever the player goes, the camera follows. A second method is to make to camera follow the transform of the player. Third method is to use the cinemachiene extention.